move get_roi and random_position to AssetGenerator
