+++ blenderset/assets.py
@@ -1,5 +1,7 @@
 import json
 from pathlib import Path
+from shapely.geometry import MultiPolygon, Polygon, Point
+import numpy as np
 
 import bpy
 
@@ -80,6 +82,29 @@ class AssetGenerator:
             if obj.get("blenderset.object_class") == class_name:
                 objs.append(obj)
         return objs
+
+    def get_roi(self):
+        if self.override_roi is not None:
+            return self.override_roi
+        walkable_polys = self.get_all_proprty_values("blenderset.visible_walkable_roi")
+        return MultiPolygon([Polygon(p) for p in walkable_polys])
+
+    def random_position(self, roi):
+        if roi.bounds:
+            min_x, min_y, max_x, max_y = roi.bounds
+            while True:
+                random_point = Point(
+                    [np.random.uniform(min_x, max_x), np.random.uniform(min_y, max_y)]
+                )
+                if random_point.within(roi):
+                    return random_point.x, random_point.y
+        else:
+            min_x, min_y, max_x, max_y = -1, -1, 1, 1
+            random_point = Point(
+                [np.random.uniform(min_x, max_x), np.random.uniform(min_y, max_y)]
+            )
+            return random_point.x, random_point.y
+
 
 
 class ComposedAssetGenerator(AssetGenerator):
